+++ src/module/patching.ts
@@ -516,7 +516,7 @@ export function _getInitiativeFormula(wrapped) {
   const original = wrapped();
   const actor = this.actor;
   if (!actor) return "1d20";
-  let disadv = actor.getFlag(game.system.id, "iniitiativeDis");
+  let disadv = actor.getFlag(game.system.id, "iniitiativeDisadv");
   let adv = actor.getFlag(game.system.id, "initiativeAdv");
   const flags = actor.data.flags["midi-qol"];
   if (flags && flags.advantage) {
@@ -577,7 +577,8 @@ export function readyPatching() {
 }
 
 export let visionPatching = () => {
-  const patchVision = isNewerVersion(game.data.version, "0.7.0") && game.settings.get("midi-qol", "playerControlsInvisibleTokens")
+  //@ts-ignore game.verison
+  const patchVision = isNewerVersion(game.version ? game.version : game.data.version, "0.7.0") && game.settings.get("midi-qol", "playerControlsInvisibleTokens")
   if (patchVision) {
     ui.notifications?.warn("Player control vision is deprecated please use the module Your Tokens Visible")
     console.warn("midi-qol | Player control vision is deprecated please use the module Your Tokens Visible")